requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.3
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
beautifulsoup4==4.12.3
lxml==5.3.0
